Reader is now subordinate to the SqlMessageStore.

+++ src/Hydrospanner/Persistence/SqlPersistence/SqlMessageStoreReader.cs
@@ -72,7 +72,7 @@
 			throw new NotSupportedException();
 		}
 
-		public SqlMessageStoreReader(DbProviderFactory factory, string connectionString, IEnumerable<string> types, int startingSequence)
+		public SqlMessageStoreReader(DbProviderFactory factory, string connectionString, IEnumerable<string> types, long startingSequence)
 		{
 			if (factory == null)
 				throw new ArgumentNullException("factory");
